@@ -1,295 +1,65 @@
 <?php
 
+declare(strict_types=1);
+
 namespace VladimirYuldashev\Flysystem;
 
 use DateTime;
-use League\Flysystem\Adapter\AbstractFtpAdapter;
-use League\Flysystem\AdapterInterface;
+use Generator;
 use League\Flysystem\Config;
-use League\Flysystem\Util;
-use League\Flysystem\Util\MimeType;
-use Normalizer;
+use League\Flysystem\DirectoryAttributes;
+use League\Flysystem\FileAttributes;
+use League\Flysystem\FilesystemAdapter;
+use League\Flysystem\PathPrefixer;
+use League\Flysystem\StorageAttributes;
+use League\Flysystem\UnableToCopyFile;
+use League\Flysystem\UnableToCreateDirectory;
+use League\Flysystem\UnableToDeleteDirectory;
+use League\Flysystem\UnableToDeleteFile;
+use League\Flysystem\UnableToMoveFile;
+use League\Flysystem\UnableToReadFile;
+use League\Flysystem\UnableToRetrieveMetadata;
+use League\Flysystem\UnableToSetVisibility;
+use League\Flysystem\UnableToWriteFile;
+use League\Flysystem\UnixVisibility\PortableVisibilityConverter;
+use League\Flysystem\UnixVisibility\VisibilityConverter;
+use League\MimeTypeDetection\FinfoMimeTypeDetector;
+use League\MimeTypeDetection\MimeTypeDetector;
 use RuntimeException;
+use Throwable;
 
-class CurlFtpAdapter extends AbstractFtpAdapter
+class CurlFtpAdapter implements FilesystemAdapter
 {
-    protected $configurable = [
-        'host',
-        'port',
-        'username',
-        'password',
-        'root',
-        'ftps',
-        'ssl',
-        'sslVerifyPeer',
-        'sslVerifyHost',
-        'utf8',
-        'timeout',
-        'passive',
-        'skipPasvIp',
-        'proxyHost',
-        'proxyPort',
-        'proxyUsername',
-        'proxyPassword',
-        'verbose',
-        'enableTimestampsOnUnixListings',
-        'useListCommandArguments',
-        'isPureFtpd',
-    ];
+    private const SYSTEM_TYPE_WINDOWS = 'windows';
+    private const SYSTEM_TYPE_UNIX = 'unix';
 
-    /** @var bool */
-    protected $useListCommandArguments = true;
+    private ?Curl $connection = null;
+    private PathPrefixer $prefixer;
+    private VisibilityConverter $visibilityConverter;
+    private ?bool $isPureFtpdServer = null;
+    private ?bool $useRawListOptions;
+    private ?string $systemType;
+    private MimeTypeDetector $mimeTypeDetector;
 
-    /** @var Curl */
-    protected $connection;
+    private ?string $rootDirectory = null;
+    protected string $separator = '/';
 
-    /** @var int unix timestamp when connection was established */
-    protected $connectionTimestamp = 0;
-
-    /** @var bool */
-    protected $isPureFtpd = null;
-
-    /** @var bool */
-    protected $ftps = true;
-
-    /** @var @int */
-    protected $sslVerifyPeer = 1;
-
-    /** @var @int */
-    protected $sslVerifyHost = 2;
-
-    /** @var bool */
-    protected $utf8 = false;
-
-    /** @var bool */
-    protected $skipPasvIp = true;
-
-    /** @var string */
-    protected $proxyHost;
-
-    /** @var int */
-    protected $proxyPort;
-
-    /** @var string */
-    protected $proxyUsername;
-
-    /** @var string */
-    protected $proxyPassword;
-
-    /** @var bool */
-    protected $verbose = false;
-
-    /**
-     * Constructor.
-     *
-     * @param  array  $config
-     */
-    public function __construct(array $config)
-    {
-        parent::__construct($config);
-
-        if (isset($config['root'])) {
-            $this->setPathPrefix($config['root']);
-        }
+    public function __construct(
+        private CurlFtpConnectionOptions $connectionOptions,
+        VisibilityConverter $visibilityConverter = null,
+        MimeTypeDetector $mimeTypeDetector = null,
+        private bool $detectMimeTypeUsingPath = false,
+    ) {
+        $this->systemType = $this->connectionOptions->systemType();
+        $this->visibilityConverter = $visibilityConverter ?: new PortableVisibilityConverter();
+        $this->mimeTypeDetector = $mimeTypeDetector ?: new FinfoMimeTypeDetector();
+        $this->useRawListOptions = $connectionOptions->useRawListOptions();
     }
 
     /**
-     * @param  bool  $ftps
+     * Disconnect FTP connection on destruct.
      */
-    public function setUseListCommandArguments($use): void
-    {
-        $this->useListCommandArguments = (bool) $use;
-    }
-
-    /**
-     * @param  bool  $ftps
-     */
-    public function setFtps($ftps): void
-    {
-        $this->ftps = (bool) $ftps;
-    }
-
-    /**
-     * @param  bool  $ssl
-     */
-    public function setSsl($ssl): void
-    {
-        $this->ssl = (bool) $ssl;
-    }
-
-    /**
-     * @param  int  $sslVerifyPeer
-     */
-    public function setSslVerifyPeer($sslVerifyPeer): void
-    {
-        $this->sslVerifyPeer = $sslVerifyPeer;
-    }
-
-    /**
-     * @param  int  $sslVerifyHost
-     */
-    public function setSslVerifyHost($sslVerifyHost): void
-    {
-        $this->sslVerifyHost = $sslVerifyHost;
-    }
-
-    /**
-     * @param  bool  $utf8
-     */
-    public function setUtf8($utf8): void
-    {
-        $this->utf8 = (bool) $utf8;
-    }
-
-    /**
-     * @param  bool  $isPureFtpd
-     */
-    public function setIsPureFtpd($isPureFtpd): void
-    {
-        $this->isPureFtpd = (bool) $isPureFtpd;
-    }
-
-    /**
-     * @param  bool  $passive
-     */
-    public function setPassive($passive): void
-    {
-        $this->passive = (bool) $passive;
-    }
-
-    /**
-     * @param  bool  $skipPasvIp
-     */
-    public function setSkipPasvIp($skipPasvIp): void
-    {
-        $this->skipPasvIp = (bool) $skipPasvIp;
-    }
-
-    /**
-     * @return string
-     */
-    public function getProxyHost()
-    {
-        return $this->proxyHost;
-    }
-
-    /**
-     * @param  string  $proxyHost
-     */
-    public function setProxyHost($proxyHost): void
-    {
-        $this->proxyHost = $proxyHost;
-    }
-
-    /**
-     * @return int
-     */
-    public function getProxyPort()
-    {
-        return $this->proxyPort;
-    }
-
-    /**
-     * @param  int  $proxyPort
-     */
-    public function setProxyPort($proxyPort): void
-    {
-        $this->proxyPort = $proxyPort;
-    }
-
-    /**
-     * @return string
-     */
-    public function getProxyUsername()
-    {
-        return $this->proxyUsername;
-    }
-
-    /**
-     * @param  string  $proxyUsername
-     */
-    public function setProxyUsername($proxyUsername): void
-    {
-        $this->proxyUsername = $proxyUsername;
-    }
-
-    /**
-     * @return string
-     */
-    public function getProxyPassword()
-    {
-        return $this->proxyPassword;
-    }
-
-    /**
-     * @param  string  $proxyPassword
-     */
-    public function setProxyPassword($proxyPassword): void
-    {
-        $this->proxyPassword = $proxyPassword;
-    }
-
-    /**
-     * @param  bool  $verbose
-     */
-    public function setVerbose($verbose): void
-    {
-        $this->verbose = (bool) $verbose;
-    }
-
-    /**
-     * Establish a connection.
-     */
-    public function connect(): void
-    {
-        $this->connection = new Curl();
-        $this->connection->setOptions([
-            CURLOPT_URL => $this->getBaseUri(),
-            CURLOPT_USERPWD => $this->getUsername().':'.$this->getPassword(),
-            CURLOPT_FTPSSLAUTH => CURLFTPAUTH_TLS,
-            CURLOPT_RETURNTRANSFER => true,
-            CURLOPT_CONNECTTIMEOUT => $this->getTimeout(),
-        ]);
-
-        if ($this->ssl) {
-            $this->connection->setOption(CURLOPT_USE_SSL, CURLFTPSSL_ALL);
-        }
-
-        if (! $this->passive) {
-            $this->connection->setOption(CURLOPT_FTPPORT, '-');
-        }
-
-        if ($this->skipPasvIp) {
-            $this->connection->setOption(CURLOPT_FTP_SKIP_PASV_IP, $this->skipPasvIp);
-        }
-
-        $this->connection->setOption(CURLOPT_SSL_VERIFYHOST, $this->sslVerifyHost);
-        $this->connection->setOption(CURLOPT_SSL_VERIFYPEER, $this->sslVerifyPeer);
-
-        if ($proxyUrl = $this->getProxyHost()) {
-            $proxyPort = $this->getProxyPort();
-            $this->connection->setOption(CURLOPT_PROXY, $proxyPort ? $proxyUrl.':'.$proxyPort : $proxyUrl);
-            $this->connection->setOption(CURLOPT_HTTPPROXYTUNNEL, true);
-        }
-
-        if ($username = $this->getProxyUsername()) {
-            $this->connection->setOption(CURLOPT_PROXYUSERPWD, $username.':'.$this->getProxyPassword());
-        }
-
-        if ($this->verbose) {
-            $this->connection->setOption(CURLOPT_VERBOSE, $this->verbose);
-        }
-
-        $this->pingConnection();
-        $this->connectionTimestamp = time();
-        $this->setUtf8Mode();
-        $this->setConnectionRoot();
-    }
-
-    /**
-     * Close the connection.
-     */
-    public function disconnect(): void
+    public function __destruct()
     {
         if ($this->connection !== null) {
             $this->connection = null;
@@ -297,543 +67,110 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     }
 
     /**
-     * Check if a connection is active.
-     *
-     * @return bool
+     * Establish a connection.
      */
-    public function isConnected()
+    private function connect(): void
     {
-        return $this->connection !== null && ! $this->hasConnectionReachedTimeout();
-    }
-
-    /**
-     * @return bool
-     */
-    protected function hasConnectionReachedTimeout()
-    {
-        return $this->connectionTimestamp + $this->getTimeout() < time();
-    }
-
-    /**
-     * Write a new file.
-     *
-     * @param  string  $path
-     * @param  string  $contents
-     * @param  Config  $config  Config object
-     * @return array|false false on failure file meta data on success
-     */
-    public function write($path, $contents, Config $config)
-    {
-        $stream = fopen('php://temp', 'w+b');
-        fwrite($stream, $contents);
-        rewind($stream);
-
-        $result = $this->writeStream($path, $stream, $config);
-
-        if ($result === false) {
-            return false;
-        }
-
-        $result['contents'] = $contents;
-        $result['mimetype'] = Util::guessMimeType($path, $contents);
-
-        return $result;
-    }
-
-    /**
-     * Write a new file using a stream.
-     *
-     * @param  string  $path
-     * @param  resource  $resource
-     * @param  Config  $config  Config object
-     * @return array|false false on failure file meta data on success
-     */
-    public function writeStream($path, $resource, Config $config)
-    {
-        $connection = $this->getConnection();
-
-        $pathDir = pathinfo($path, PATHINFO_DIRNAME);
-        $pathHasFolders = $pathDir !== '.';
-        $requestPath = $pathHasFolders ? $this->applyPathPrefix($path) : $path;
-
-        $result = $connection->exec([
-            CURLOPT_URL => $this->getBaseUri().$this->separator.rawurlencode($requestPath),
-            CURLOPT_UPLOAD => 1,
-            CURLOPT_INFILE => $resource,
+        $this->connection = new Curl();
+        $this->connection->setOptions([
+            CURLOPT_URL => $this->connectionOptions->baseUrl(),
+            CURLOPT_USERPWD => $this->connectionOptions->username().':'.$this->connectionOptions->password(),
+            CURLOPT_FTPSSLAUTH => CURLFTPAUTH_TLS,
+            CURLOPT_RETURNTRANSFER => true,
+            CURLOPT_CONNECTTIMEOUT => $this->connectionOptions->timeout(),
         ]);
 
-        if ($result === false) {
-            return false;
+        if ($this->connectionOptions->ssl()) {
+            $this->connection->setOption(CURLOPT_USE_SSL, CURLFTPSSL_ALL);
         }
 
-        if ($pathHasFolders) {
-            $this->setConnectionRoot();
+        if (! $this->connectionOptions->passive()) {
+            $this->connection->setOption(CURLOPT_FTPPORT, '-');
         }
 
-        $type = 'file';
+        if ($this->connectionOptions->ignorePassiveAddress() != null) {
+            $this->connection->setOption(CURLOPT_FTP_SKIP_PASV_IP, $this->connectionOptions->ignorePassiveAddress());
+        }
 
-        return compact('type', 'path');
+        $this->connection->setOption(CURLOPT_SSL_VERIFYHOST, $this->connectionOptions->sslVerifyHost());
+        $this->connection->setOption(CURLOPT_SSL_VERIFYPEER, $this->connectionOptions->sslVerifyPeer());
+
+        if ($proxyUrl = $this->connectionOptions->proxyHost()) {
+            $proxyPort = $this->connectionOptions->proxyPort();
+            $this->connection->setOption(CURLOPT_PROXY, $proxyPort ? $proxyUrl.':'.$proxyPort : $proxyUrl);
+            $this->connection->setOption(CURLOPT_HTTPPROXYTUNNEL, true);
+        }
+
+        if ($username = $this->connectionOptions->proxyUsername()) {
+            $this->connection->setOption(CURLOPT_PROXYUSERPWD, $username.':'.$this->connectionOptions->proxyPassword());
+        }
+
+        if ($this->connectionOptions->verbose()) {
+            $this->connection->setOption(CURLOPT_VERBOSE, $this->connectionOptions->verbose());
+        }
+
+        $this->pingConnection();
+        $this->setUtf8Mode();
+        $this->setConnectionRoot();
     }
 
     /**
-     * Update a file.
-     *
-     * @param  string  $path
-     * @param  string  $contents
-     * @param  Config  $config  Config object
-     * @return array|false false on failure file meta data on success
+     * Check the connection is established.
      */
-    public function update($path, $contents, Config $config)
+    protected function pingConnection(): void
     {
-        return $this->write($path, $contents, $config);
+        if ($this->connection->exec() === false) {
+            throw new RuntimeException('Could not connect to host: '.$this->connectionOptions->host().', port:'.$this->connectionOptions->port().'. Error : '.$this->connection->lastError());
+        }
+    }
+
+    protected function isConnected(): bool
+    {
+        return $this->connection !== null && $this->connection->exec() !== false;
     }
 
     /**
-     * Update a file using a stream.
-     *
-     * @param  string  $path
-     * @param  resource  $resource
-     * @param  Config  $config  Config object
-     * @return array|false false on failure file meta data on success
+     * Set the connection to UTF-8 mode.
      */
-    public function updateStream($path, $resource, Config $config)
+    protected function setUtf8Mode(): void
     {
-        return $this->writeStream($path, $resource, $config);
-    }
+        if (! $this->connectionOptions->utf8()) {
+            return;
+        }
 
-    /**
-     * Rename a file.
-     *
-     * @param  string  $path
-     * @param  string  $newpath
-     * @return bool
-     */
-    public function rename($path, $newpath)
-    {
-        $connection = $this->getConnection();
-
-        $moveCommands = [
-            'RNFR '.$path,
-            'RNTO '.$newpath,
-        ];
-
-        $response = $this->rawPost($connection, $moveCommands);
+        $response = $this->rawCommand($this->connection, 'OPTS UTF8 ON');
         [$code] = explode(' ', end($response), 2);
-
-        return (int) $code === 250;
+        if (! in_array($code, ['200', '202'])) {
+            throw new RuntimeException(
+                'Could not set UTF-8 mode for connection: '.$this->connectionOptions->host().'::'.$this->connectionOptions->port()
+            );
+        }
     }
 
     /**
-     * Copy a file.
-     *
-     * @param  string  $path
-     * @param  string  $newpath
-     * @return bool
+     * Set the connection root.
      */
-    public function copy($path, $newpath)
+    protected function setConnectionRoot(): void
     {
-        $file = $this->read($path);
-
-        if ($file === false) {
-            return false;
+        $this->rootDirectory = $root = $this->connectionOptions->root();
+        if (empty($root)) {
+            return;
         }
 
-        return $this->write($newpath, $file['contents'], new Config()) !== false;
-    }
-
-    /**
-     * Delete a file.
-     *
-     * @param  string  $path
-     * @return bool
-     */
-    public function delete($path)
-    {
-        $connection = $this->getConnection();
-
-        $response = $this->rawCommand($connection, 'DELE '.$path);
+        $response = $this->rawCommand($this->connection, 'CWD '.$root);
         [$code] = explode(' ', end($response), 2);
-
-        return (int) $code === 250;
-    }
-
-    /**
-     * Delete a directory.
-     *
-     * @param  string  $dirname
-     * @return bool
-     */
-    public function deleteDir($dirname)
-    {
-        $connection = $this->getConnection();
-
-        $response = $this->rawCommand($connection, 'RMD '.$dirname);
-        [$code] = explode(' ', end($response), 2);
-
-        return (int) $code === 250;
-    }
-
-    /**
-     * Create a directory.
-     *
-     * @param  string  $dirname  directory name
-     * @param  Config  $config
-     * @return array|false
-     */
-    public function createDir($dirname, Config $config)
-    {
-        $connection = $this->getConnection();
-
-        $response = $this->rawCommand($connection, 'MKD '.$dirname);
-        [$code] = explode(' ', end($response), 2);
-        if ((int) $code !== 257) {
-            return false;
+        if ((int) $code !== 250) {
+            throw new RuntimeException('Unable to resolve connection root. It does not seem to exist: '.$root);
         }
-
-        return ['type' => 'dir', 'path' => $dirname];
-    }
-
-    /**
-     * Set the visibility for a file.
-     *
-     * @param  string  $path
-     * @param  string  $visibility
-     * @return array|false file meta data
-     */
-    public function setVisibility($path, $visibility)
-    {
-        $connection = $this->getConnection();
-
-        if ($visibility === AdapterInterface::VISIBILITY_PUBLIC) {
-            $mode = $this->getPermPublic();
-        } else {
-            $mode = $this->getPermPrivate();
-        }
-
-        $request = sprintf('SITE CHMOD %o %s', $mode, $path);
-        $response = $this->rawCommand($connection, $request);
-        [$code] = explode(' ', end($response), 2);
-        if ((int) $code !== 200) {
-            return false;
-        }
-
-        return $this->getMetadata($path);
-    }
-
-    /**
-     * Read a file.
-     *
-     * @param  string  $path
-     * @return array|false
-     */
-    public function read($path)
-    {
-        if (! $object = $this->readStream($path)) {
-            return false;
-        }
-
-        $object['contents'] = stream_get_contents($object['stream']);
-        fclose($object['stream']);
-        unset($object['stream']);
-
-        return $object;
-    }
-
-    /**
-     * Read a file as a stream.
-     *
-     * @param  string  $path
-     * @return array|false
-     */
-    public function readStream($path)
-    {
-        $stream = fopen('php://temp', 'w+b');
-
-        $connection = $this->getConnection();
-
-        $pathDir = pathinfo($path, PATHINFO_DIRNAME);
-        $pathHasFolders = $pathDir !== '.';
-        $requestPath = $pathHasFolders ? $this->applyPathPrefix($path) : $path;
-
-        $result = $connection->exec([
-            CURLOPT_URL => $this->getBaseUri().$this->separator.rawurlencode($requestPath),
-            CURLOPT_FILE => $stream,
-        ]);
-
-        if (! $result) {
-            fclose($stream);
-
-            return false;
-        }
-
-        rewind($stream);
-
-        if ($pathHasFolders) {
-            $this->setConnectionRoot();
-        }
-
-        return ['type' => 'file', 'path' => $path, 'stream' => $stream];
-    }
-
-    /**
-     * Get all the meta data of a file or directory.
-     *
-     * @param  string  $path
-     * @return array|false
-     */
-    public function getMetadata($path)
-    {
-        if ($path === '') {
-            return ['type' => 'dir', 'path' => ''];
-        }
-
-        $arguments = $this->useListCommandArguments ? '-A ' : '';
-        $request = rtrim('LIST '.$arguments.$this->normalizePath($path));
-
-        $connection = $this->getConnection();
-        $result = $connection->exec([CURLOPT_CUSTOMREQUEST => $request]);
-        if ($result === false) {
-            return false;
-        }
-        $listing = $this->normalizeListing(explode(PHP_EOL, $result), '');
-
-        $pathIsDir = pathinfo($path, PATHINFO_EXTENSION) === '';
-
-        if ($pathIsDir && count($listing) === 0) {
-            return ['type' => 'dir', 'path' => $path];
-        }
-
-        return current($listing);
-    }
-
-    /**
-     * Normalize a Unix file entry.
-     *
-     * Given $item contains:
-     *    '-rw-r--r--   1 ftp      ftp           409 Aug 19 09:01 file1.txt'
-     *
-     * This function will return:
-     * [
-     *   'type' => 'file',
-     *   'path' => 'file1.txt',
-     *   'visibility' => 'public',
-     *   'size' => 409,
-     *   'timestamp' => 1566205260
-     * ]
-     *
-     * @param  string  $item
-     * @param  string  $base
-     * @return array normalized file array
-     */
-    protected function normalizeUnixObject($item, $base)
-    {
-        $item = preg_replace('#\s+#', ' ', trim($item), 7);
-
-        if (count(explode(' ', $item, 9)) !== 9) {
-            throw new RuntimeException("Metadata can't be parsed from item '$item' , not enough parts.");
-        }
-
-        [$permissions, /* $number */, /* $owner */, /* $group */, $size, $month, $day, $timeOrYear, $name] = explode(' ', $item, 9);
-        $type = $this->detectType($permissions);
-        $path = $base === '' ? $name : $base.$this->separator.$name;
-
-        if ($type === 'dir') {
-            $result = compact('type', 'path');
-            if ($this->enableTimestampsOnUnixListings) {
-                $timestamp = $this->normalizeUnixTimestamp($month, $day, $timeOrYear);
-                $result += compact('timestamp');
-            }
-
-            return $result;
-        }
-
-        $pathPrefix = $this->getPathPrefix();
-        if (substr($path, 0, strlen($pathPrefix)) === $pathPrefix) {
-            $path = $this->removePathPrefix($path);
-        }
-
-        $permissions = $this->normalizePermissions($permissions);
-        $visibility = $permissions & 0044 ? AdapterInterface::VISIBILITY_PUBLIC : AdapterInterface::VISIBILITY_PRIVATE;
-        $size = (int) $size;
-
-        $result = compact('type', 'path', 'visibility', 'size');
-        if ($this->enableTimestampsOnUnixListings) {
-            $timestamp = $this->normalizeUnixTimestamp($month, $day, $timeOrYear);
-            $result += compact('timestamp');
-        }
-
-        return $result;
-    }
-
-    /**
-     * Get the mimetype of a file.
-     *
-     * @param  string  $path
-     * @return array|false
-     */
-    public function getMimetype($path)
-    {
-        if (! $metadata = $this->getMetadata($path)) {
-            return false;
-        }
-
-        $metadata['mimetype'] = MimeType::detectByFilename($path);
-
-        return $metadata;
-    }
-
-    /**
-     * Get the timestamp of a file.
-     *
-     * @param  string  $path
-     * @return array|false
-     */
-    public function getTimestamp($path)
-    {
-        $response = $this->rawCommand($this->getConnection(), 'MDTM '.$path);
-        [$code, $time] = explode(' ', end($response), 2);
-        if ($code !== '213') {
-            return false;
-        }
-
-        if (strpos($time, '.')) {
-            $datetime = DateTime::createFromFormat('YmdHis.u', $time);
-        } else {
-            $datetime = DateTime::createFromFormat('YmdHis', $time);
-        }
-
-        if (! $datetime) {
-            return false;
-        }
-
-        return ['path' => $path, 'timestamp' => $datetime->getTimestamp()];
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @param  string  $directory
-     */
-    protected function listDirectoryContents($directory, $recursive = false)
-    {
-        if ($recursive === true) {
-            return $this->listDirectoryContentsRecursive($directory);
-        }
-
-        $arguments = $this->useListCommandArguments ? '-aln ' : '';
-        $request = rtrim('LIST '.$arguments.$this->normalizePath($directory));
-
-        $connection = $this->getConnection();
-        $result = $connection->exec([CURLOPT_CUSTOMREQUEST => $request]);
-        if ($result === false) {
-            return [];
-        }
-
-        if ($directory === '/') {
-            $directory = '';
-        }
-
-        return $this->normalizeListing(explode(PHP_EOL, $result), $directory);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @param  string  $directory
-     */
-    protected function listDirectoryContentsRecursive($directory)
-    {
-        $arguments = $this->useListCommandArguments ? '-aln ' : '';
-        $request = rtrim('LIST '.$arguments.$this->normalizePath($directory));
-
-        $connection = $this->getConnection();
-        $result = $connection->exec([CURLOPT_CUSTOMREQUEST => $request]);
-
-        $listing = $this->normalizeListing(explode(PHP_EOL, $result), $directory);
-        $output = [];
-
-        foreach ($listing as $item) {
-            $output[] = $item;
-            if ($item['type'] === 'dir') {
-                $output = array_merge($output, $this->listDirectoryContentsRecursive($item['path']));
-            }
-        }
-
-        return $output;
-    }
-
-    /**
-     * Normalize a permissions string.
-     *
-     * @param  string  $permissions
-     * @return int
-     */
-    protected function normalizePermissions($permissions)
-    {
-        // remove the type identifier
-        $permissions = substr($permissions, 1);
-        // map the string rights to the numeric counterparts
-        $map = ['-' => '0', 'r' => '4', 'w' => '2', 'x' => '1'];
-        $permissions = strtr($permissions, $map);
-        // split up the permission groups
-        $parts = str_split($permissions, 3);
-        // convert the groups
-        $mapper = function ($part) {
-            return array_sum(str_split($part));
-        };
-
-        // converts to decimal number
-        return octdec(implode('', array_map($mapper, $parts)));
-    }
-
-    /**
-     * Normalize path depending on server.
-     *
-     * @param  string  $path
-     * @return string
-     */
-    protected function normalizePath($path)
-    {
-        if (empty($path)) {
-            return '';
-        }
-
-        $path = Normalizer::normalize($path);
-
-        if ($this->isPureFtpdServer()) {
-            $path = str_replace(' ', '\ ', $path);
-        }
-
-        $path = str_replace('*', '\\*', $path);
-
-        return $path;
-    }
-
-    /**
-     * @return bool
-     */
-    protected function isPureFtpdServer()
-    {
-        if ($this->isPureFtpd === null) {
-            $response = $this->rawCommand($this->getConnection(), 'HELP');
-            $response = end($response);
-            $this->isPureFtpd = stripos($response, 'Pure-FTPd') !== false;
-        }
-
-        return $this->isPureFtpd;
     }
 
     /**
      * Sends an arbitrary command to an FTP server.
-     *
-     * @param  Curl  $connection  The CURL instance
-     * @param  string  $command  The command to execute
-     * @return array Returns the server's response as an array of strings
      */
-    protected function rawCommand($connection, $command)
+    protected function rawCommand(Curl $connection, string $command): array
     {
         $response = '';
-        $callback = static function ($ch, $string) use (&$response) {
+        $callback = static function ($curl, $string) use (&$response) {
             $response .= $string;
 
             return strlen($string);
@@ -849,15 +186,11 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * Sends an arbitrary command to an FTP server using POSTQUOTE option. This makes sure all commands are run
      * in succession and increases chance of success for complex operations like "move/rename file".
-     *
-     * @param  Curl  $connection  The CURL instance
-     * @param  array  $commandsArray  The commands to execute
-     * @return array Returns the server's response as an array of strings
      */
-    protected function rawPost($connection, array $commandsArray)
+    protected function rawPost(Curl $connection, array $commandsArray): array
     {
         $response = '';
-        $callback = function ($ch, $string) use (&$response) {
+        $callback = function ($curl, $string) use (&$response) {
             $response .= $string;
 
             return strlen($string);
@@ -870,62 +203,598 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         return explode(PHP_EOL, trim($response));
     }
 
-    /**
-     * Returns the base url of the connection.
-     *
-     * @return string
-     */
-    protected function getBaseUri()
+    private function connection(): ?Curl
     {
-        $protocol = $this->ftps ? 'ftps' : 'ftp';
+        start:
+        if ($this->connection === null) {
+            $this->connect();
+            $this->prefixer = new PathPrefixer($this->rootDirectory);
 
-        return $protocol.'://'.$this->getHost().':'.$this->getPort();
+            return $this->connection;
+        }
+
+        if (! $this->isConnected()) {
+            $this->connection = null;
+            goto start;
+        }
+
+        $this->setConnectionRoot();
+
+        return $this->connection;
     }
 
-    /**
-     * Check the connection is established.
-     */
-    protected function pingConnection(): void
+    private function isPureFtpdServer(): bool
     {
-        // We can't use the getConnection, because it will lead to an infinite cycle
-        if ($this->connection->exec() === false) {
-            throw new RuntimeException('Could not connect to host: '.$this->getHost().', port:'.$this->getPort());
+        if ($this->isPureFtpdServer !== null) {
+            return $this->isPureFtpdServer;
+        }
+
+        $response = $this->rawCommand($this->connection(), 'HELP');
+        $response = end($response);
+        $this->isPureFtpdServer = stripos($response, 'Pure-FTPd') !== false;
+
+        return $this->isPureFtpdServer;
+    }
+
+    private function isServerSupportingListOptions(): bool
+    {
+        if ($this->useRawListOptions !== null) {
+            return $this->useRawListOptions;
+        }
+
+        $response = $this->rawCommand($this->connection(), 'SYST');
+        $syst = implode(' ', $response);
+
+        return $this->useRawListOptions = stripos($syst, 'FileZilla') === false
+            && stripos($syst, 'L8') === false;
+    }
+
+    public function fileExists(string $path): bool
+    {
+        try {
+            $this->fileSize($path);
+
+            return true;
+        } catch (UnableToRetrieveMetadata $exception) {
+            return false;
         }
     }
 
-    /**
-     * Set the connection to UTF-8 mode.
-     */
-    protected function setUtf8Mode(): void
+    public function write(string $path, string $contents, Config $config): void
     {
-        if (! $this->utf8) {
+        try {
+            $writeStream = fopen('php://temp', 'w+b');
+            fwrite($writeStream, $contents);
+            rewind($writeStream);
+            $this->writeStream($path, $writeStream, $config);
+        } finally {
+            isset($writeStream) && is_resource($writeStream) && fclose($writeStream);
+        }
+    }
+
+    public function writeStream(string $path, $contents, Config $config): void
+    {
+        try {
+            $this->ensureParentDirectoryExists($path, $config->get(Config::OPTION_DIRECTORY_VISIBILITY));
+        } catch (Throwable $exception) {
+            throw UnableToWriteFile::atLocation($path, 'creating parent directory failed', $exception);
+        }
+
+        $location = $this->prefixer()->prefixPath($path);
+
+        $result = $this->connection()->exec([
+            CURLOPT_URL => $this->connectionOptions->baseUrl().$this->separator.rawurlencode($location),
+            CURLOPT_UPLOAD => 1,
+            CURLOPT_INFILE => $contents,
+        ]);
+
+        if ($result === false) {
+            throw UnableToWriteFile::atLocation($path, $this->connection->lastError());
+        }
+
+        if (! $visibility = $config->get(Config::OPTION_VISIBILITY)) {
             return;
         }
 
-        $response = $this->rawCommand($this->connection, 'OPTS UTF8 ON');
-        [$code, $message] = explode(' ', end($response), 2);
-        if (! in_array($code, ['200', '202'])) {
-            throw new RuntimeException(
-                'Could not set UTF-8 mode for connection: '.$this->getHost().'::'.$this->getPort()
+        try {
+            $this->setVisibility($path, (string) $visibility);
+        } catch (Throwable $exception) {
+            throw UnableToWriteFile::atLocation($path, 'setting visibility failed', $exception);
+        }
+    }
+
+    public function read(string $path): string
+    {
+        $readStream = $this->readStream($path);
+        $contents = stream_get_contents($readStream);
+        fclose($readStream);
+
+        return $contents;
+    }
+
+    public function readStream(string $path)
+    {
+        $location = $this->prefixer()->prefixPath($path);
+        $stream = fopen('php://temp', 'w+b');
+        $result = $this->connection()->exec([
+            CURLOPT_URL => $this->connectionOptions->baseUrl().$this->separator.rawurlencode($location),
+            CURLOPT_FILE => $stream,
+        ]);
+
+        if (! $result) {
+            fclose($stream);
+
+            throw UnableToReadFile::fromLocation($path, $this->connection->lastError());
+        }
+
+        rewind($stream);
+
+        return $stream;
+    }
+
+    public function delete(string $path): void
+    {
+        $this->deleteFile($path);
+    }
+
+    private function deleteFile(string $path): void
+    {
+        $location = $this->prefixer()->prefixPath($path);
+        $connection = $this->connection();
+
+        $response = $this->rawCommand($connection, 'DELE '.$location);
+        [$code] = explode(' ', end($response), 2);
+
+        if ((int) $code !== 250) {
+            throw UnableToDeleteFile::atLocation($path, 'the file still exists');
+        }
+    }
+
+    public function deleteDirectory(string $path): void
+    {
+        /** @var StorageAttributes[] $contents */
+        $contents = $this->listContents($path, true);
+        $connection = $this->connection();
+        $directories = [$path];
+
+        foreach ($contents as $item) {
+            if ($item->isDir()) {
+                $directories[] = $item->path();
+                continue;
+            }
+            try {
+                $this->deleteFile($item->path());
+            } catch (Throwable $exception) {
+                throw UnableToDeleteDirectory::atLocation($path, 'unable to delete child', $exception);
+            }
+        }
+
+        rsort($directories);
+
+        foreach ($directories as $directory) {
+            $directoryLocation = $this->prefixer()->prefixPath($directory);
+            $response = $this->rawCommand($connection, 'RMD '.$directoryLocation);
+            [$code] = explode(' ', end($response), 2);
+            if ((int) $code !== 250) {
+                throw UnableToDeleteDirectory::atLocation($path, "Could not delete directory $directory");
+            }
+        }
+    }
+
+    public function createDirectory(string $path, Config $config): void
+    {
+        $this->ensureDirectoryExists($path, $config->get('directory_visibility', $config->get('visibility')));
+    }
+
+    public function setVisibility(string $path, string $visibility): void
+    {
+        $location = $this->prefixer()->prefixPath($path);
+        $mode = $this->visibilityConverter->forFile($visibility);
+
+        $connection = $this->connection();
+        $request = sprintf('SITE CHMOD %o %s', $mode, $location);
+        $response = $this->rawCommand($connection, $request);
+        [$code, $errorMessage] = explode(' ', end($response), 2);
+        if ((int) $code !== 200) {
+            $errorMessage = 'unable to chmod the file by running SITE CHMOD: '.$errorMessage;
+            throw UnableToSetVisibility::atLocation($path, $errorMessage);
+        }
+    }
+
+    private function fetchMetadata(string $path, string $type): FileAttributes
+    {
+        $location = $this->prefixer()->prefixPath($path);
+
+        if ($this->isPureFtpdServer()) {
+            $location = str_replace(' ', '\ ', $location);
+            $location = $this->escapePath($location);
+        }
+
+        $connection = $this->connection();
+        $object = $this->rawCommand($connection, 'STAT '.$location);
+
+        if (empty($object) || count($object) < 4 || substr($object[1], 0, 5) === 'ftpd:') {
+            throw UnableToRetrieveMetadata::create($path, $type, $this->connection->lastError());
+        }
+
+        $attributes = $this->normalizeObject($object[2], '');
+
+        if (! $attributes instanceof FileAttributes) {
+            throw UnableToRetrieveMetadata::create(
+                $path,
+                $type,
+                'expected file, '.($attributes instanceof DirectoryAttributes ? 'directory found' : 'nothing found')
             );
         }
+
+        return $attributes;
     }
 
-    /**
-     * Set the connection root.
-     */
-    protected function setConnectionRoot(): void
+    public function mimeType(string $path): FileAttributes
     {
-        $root = $this->getRoot();
-        if (empty($root)) {
+        try {
+            $mimetype = $this->detectMimeTypeUsingPath
+                ? $this->mimeTypeDetector->detectMimeTypeFromPath($path)
+                : $this->mimeTypeDetector->detectMimeType($path, $this->read($path));
+        } catch (Throwable $exception) {
+            throw UnableToRetrieveMetadata::mimeType($path, $exception->getMessage(), $exception);
+        }
+
+        if ($mimetype === null) {
+            throw UnableToRetrieveMetadata::mimeType($path, 'Unknown.');
+        }
+
+        return new FileAttributes($path, null, null, null, $mimetype);
+    }
+
+    public function lastModified(string $path): FileAttributes
+    {
+        $location = $this->prefixer()->prefixPath($path);
+        $connection = $this->connection();
+
+        $response = $this->rawCommand($connection, 'MDTM '.$location);
+        [$code, $time] = explode(' ', end($response), 2);
+        if ($code !== '213') {
+            throw UnableToRetrieveMetadata::lastModified($path, $time);
+        }
+
+        if (strpos($time, '.')) {
+            $lastModified = DateTime::createFromFormat('YmdHis.u', $time);
+        } else {
+            $lastModified = DateTime::createFromFormat('YmdHis', $time);
+        }
+
+        if (! $lastModified) {
+            throw UnableToRetrieveMetadata::lastModified($path, "Can't parse lastModified string {$time} to timestamp");
+        }
+
+        return new FileAttributes($path, null, null, $lastModified->getTimestamp());
+    }
+
+    public function visibility(string $path): FileAttributes
+    {
+        return $this->fetchMetadata($path, FileAttributes::ATTRIBUTE_VISIBILITY);
+    }
+
+    public function fileSize(string $path): FileAttributes
+    {
+        $location = $this->prefixer()->prefixPath($path);
+        $response = $this->rawCommand($this->connection(), 'SIZE '.$location);
+        [$code, $fileSize] = explode(' ', end($response), 2);
+        if ($code != '213') {
+            throw UnableToRetrieveMetadata::fileSize($path, $fileSize);
+        }
+
+        if ($fileSize < 0) {
+            throw UnableToRetrieveMetadata::fileSize($path, 'Filesize is less then 0');
+        }
+
+        return new FileAttributes($path, (int) $fileSize);
+    }
+
+    private function ftpRawlist(string $options, string $path): array
+    {
+        $path = rtrim($path, '/').'/';
+
+        if ($this->isPureFtpdServer()) {
+            $path = str_replace(' ', '\ ', $path);
+            $path = $this->escapePath($path);
+        }
+
+        if (! $this->isServerSupportingListOptions()) {
+            $options = '';
+        }
+
+        $request = rtrim('LIST '.$options.$path);
+        $listing = $this->connection()->exec([CURLOPT_CUSTOMREQUEST => $request]);
+
+        return explode(PHP_EOL, $listing);
+    }
+
+    public function listContents(string $path, bool $deep): iterable
+    {
+        $path = ltrim($path, '/');
+        $path = $path === '' ? $path : trim($path, '/').'/';
+
+        if ($deep && $this->connectionOptions->recurseManually()) {
+            yield from $this->listDirectoryContentsRecursive($path);
+        } else {
+            $location = $this->prefixer()->prefixPath($path);
+            $options = $deep ? '-alnR' : '-aln';
+
+            $listing = $this->ftpRawlist($options, $location);
+
+            yield from $this->normalizeListing($listing, $path);
+        }
+    }
+
+    private function normalizeListing(array $listing, string $prefix = ''): Generator
+    {
+        $base = $prefix;
+
+        foreach ($listing as $item) {
+            if ($item === '' || preg_match('#.* \.(\.)?$|^total#', $item)) {
+                continue;
+            }
+
+            if (preg_match('#^.*:$#', $item)) {
+                $base = preg_replace('~^\./*|:$~', '', $item);
+                continue;
+            }
+
+            yield $this->normalizeObject($item, $base);
+        }
+    }
+
+    private function normalizeObject(string $item, string $base): StorageAttributes
+    {
+        $this->systemType === null && $this->systemType = $this->detectSystemType($item);
+
+        if ($this->systemType === self::SYSTEM_TYPE_UNIX) {
+            return $this->normalizeUnixObject($item, $base);
+        }
+
+        return $this->normalizeWindowsObject($item, $base);
+    }
+
+    private function detectSystemType(string $item): string
+    {
+        return preg_match(
+            '/^[0-9]{2,4}-[0-9]{2}-[0-9]{2}/',
+            $item
+        ) ? self::SYSTEM_TYPE_WINDOWS : self::SYSTEM_TYPE_UNIX;
+    }
+
+    private function normalizeWindowsObject(string $item, string $base): StorageAttributes
+    {
+        $item = preg_replace('#\s+#', ' ', trim($item), 3);
+        $parts = explode(' ', $item, 4);
+
+        if (count($parts) !== 4) {
+            throw new RuntimeException("Metadata can't be parsed from item '$item' , not enough parts.");
+        }
+
+        [$date, $time, $size, $name] = $parts;
+        $path = $base === '' ? $name : rtrim($base, '/').'/'.$name;
+
+        if ($size === '<DIR>') {
+            return new DirectoryAttributes($path);
+        }
+
+        // Check for the correct date/time format
+        $format = strlen($date) === 8 ? 'm-d-yH:iA' : 'Y-m-dH:i';
+        $dateTime = DateTime::createFromFormat($format, $date.$time);
+        $lastModified = $dateTime ? $dateTime->getTimestamp() : (int) strtotime("$date $time");
+
+        return new FileAttributes($path, (int) $size, null, $lastModified);
+    }
+
+    private function normalizeUnixObject(string $item, string $base): StorageAttributes
+    {
+        $item = preg_replace('#\s+#', ' ', trim($item), 7);
+        $parts = explode(' ', $item, 9);
+
+        if (count($parts) !== 9) {
+            throw new RuntimeException("Metadata can't be parsed from item '$item' , not enough parts.");
+        }
+
+        [$permissions, /* $number */, /* $owner */, /* $group */, $size, $month, $day, $timeOrYear, $name] = $parts;
+        $isDirectory = $this->listingItemIsDirectory($permissions);
+        $permissions = $this->normalizePermissions($permissions);
+        $path = $base === '' ? $name : rtrim($base, '/').'/'.$name;
+        $lastModified = $this->connectionOptions->timestampsOnUnixListingsEnabled() ? $this->normalizeUnixTimestamp(
+            $month,
+            $day,
+            $timeOrYear
+        ) : null;
+
+        if ($isDirectory) {
+            return new DirectoryAttributes(
+                $path,
+                $this->visibilityConverter->inverseForDirectory($permissions),
+                $lastModified
+            );
+        }
+
+        $visibility = $this->visibilityConverter->inverseForFile($permissions);
+
+        return new FileAttributes($path, (int) $size, $visibility, $lastModified);
+    }
+
+    private function listingItemIsDirectory(string $permissions): bool
+    {
+        return substr($permissions, 0, 1) === 'd';
+    }
+
+    private function normalizeUnixTimestamp(string $month, string $day, string $timeOrYear): int
+    {
+        if (is_numeric($timeOrYear)) {
+            $year = $timeOrYear;
+            $hour = '00';
+            $minute = '00';
+            $seconds = '00';
+        } else {
+            $year = date('Y');
+            [$hour, $minute] = explode(':', $timeOrYear);
+            $seconds = '00';
+        }
+
+        $dateTime = DateTime::createFromFormat('Y-M-j-G:i:s', "{$year}-{$month}-{$day}-{$hour}:{$minute}:{$seconds}");
+
+        return $dateTime->getTimestamp();
+    }
+
+    private function normalizePermissions(string $permissions): int
+    {
+        // remove the type identifier
+        $permissions = substr($permissions, 1);
+
+        // map the string rights to the numeric counterparts
+        $map = ['-' => '0', 'r' => '4', 'w' => '2', 'x' => '1'];
+        $permissions = strtr($permissions, $map);
+
+        // split up the permission groups
+        $parts = str_split($permissions, 3);
+
+        // convert the groups
+        $mapper = function ($part) {
+            return array_sum(str_split($part));
+        };
+
+        // converts to decimal number
+        return octdec(implode('', array_map($mapper, $parts)));
+    }
+
+    private function listDirectoryContentsRecursive(string $directory): Generator
+    {
+        $location = $this->prefixer()->prefixPath($directory);
+        $listing = $this->ftpRawlist('-aln', $location);
+        /** @var StorageAttributes[] $listing */
+        $listing = $this->normalizeListing($listing, $directory);
+
+        foreach ($listing as $item) {
+            yield $item;
+
+            if (! $item->isDir()) {
+                continue;
+            }
+
+            $children = $this->listDirectoryContentsRecursive($item->path());
+
+            foreach ($children as $child) {
+                yield $child;
+            }
+        }
+    }
+
+    public function move(string $source, string $destination, Config $config): void
+    {
+        try {
+            $this->ensureParentDirectoryExists($destination, $config->get(Config::OPTION_DIRECTORY_VISIBILITY));
+        } catch (Throwable $exception) {
+            throw UnableToMoveFile::fromLocationTo($source, $destination, $exception);
+        }
+
+        $sourceLocation = $this->prefixer()->prefixPath($source);
+        $destinationLocation = $this->prefixer()->prefixPath($destination);
+        $connection = $this->connection();
+
+        $moveCommands = [
+            'RNFR '.$sourceLocation,
+            'RNTO '.$destinationLocation,
+        ];
+
+        $response = $this->rawPost($connection, $moveCommands);
+        [$code] = explode(' ', end($response), 2);
+
+        if ((int) $code !== 250) {
+            throw UnableToMoveFile::fromLocationTo($source, $destination);
+        }
+    }
+
+    public function copy(string $source, string $destination, Config $config): void
+    {
+        try {
+            $readStream = $this->readStream($source);
+            $visibility = $this->visibility($source)->visibility();
+            $this->writeStream($destination, $readStream, new Config(compact('visibility')));
+        } catch (Throwable $exception) {
+            if (isset($readStream) && is_resource($readStream)) {
+                @fclose($readStream);
+            }
+            throw UnableToCopyFile::fromLocationTo($source, $destination, $exception);
+        }
+    }
+
+    private function ensureParentDirectoryExists(string $path, ?string $visibility): void
+    {
+        $dirname = dirname($path);
+
+        if ($dirname === '' || $dirname === '.') {
             return;
         }
 
-        // We can't use the getConnection, because it will lead to an infinite cycle
-        $response = $this->rawCommand($this->connection, 'CWD '.$root);
-        [$code] = explode(' ', end($response), 2);
-        if ((int) $code !== 250) {
-            throw new RuntimeException('Root is invalid or does not exist: '.$this->getRoot());
+        $this->ensureDirectoryExists($dirname, $visibility);
+    }
+
+    private function ensureDirectoryExists(string $dirname, ?string $visibility): void
+    {
+        $connection = $this->connection();
+
+        $dirPath = '';
+        $parts = explode('/', trim($dirname, '/'));
+        $mode = $visibility ? $this->visibilityConverter->forDirectory($visibility) : false;
+
+        foreach ($parts as $part) {
+            $dirPath .= '/'.$part;
+            $location = $this->prefixer()->prefixPath($dirPath);
+
+            if ($this->directoryExists($location)) {
+                continue;
+            }
+
+            $response = $this->rawCommand($connection, 'MKD '.$location);
+            [$code] = explode(' ', end($response), 2);
+
+            if ((int) $code !== 257) {
+                $errorMessage = 'unable to create the directory: '.$this->connection->lastError();
+                throw UnableToCreateDirectory::atLocation($dirPath, $errorMessage);
+            }
+
+            if ($mode !== false) {
+                $request = sprintf('SITE CHMOD %o %s', $mode, $location);
+                $response = $this->rawCommand($connection, $request);
+                [$code, $errorMessage] = explode(' ', end($response), 2);
+                if ((int) $code !== 200) {
+                    $errorMessage = 'unable to chmod the directory by running SITE CHMOD: '.$errorMessage;
+                    throw UnableToCreateDirectory::atLocation($dirPath, $errorMessage);
+                }
+            }
         }
+    }
+
+    private function escapePath(string $path): string
+    {
+        return str_replace(['*', '[', ']'], ['\\*', '\\[', '\\]'], $path);
+    }
+
+    public function directoryExists(string $path): bool
+    {
+        $response = $this->rawCommand($this->connection(), 'CWD '.$path);
+        [$code] = explode(' ', end($response), 2);
+
+        return (int) $code === 250;
+    }
+
+    /**
+     * @return PathPrefixer
+     */
+    private function prefixer(): PathPrefixer
+    {
+        if ($this->rootDirectory === null) {
+            $this->connection();
+        }
+
+        return $this->prefixer;
     }
 }
