@@ -1,58 +1,46 @@
 <?php
 
+declare(strict_types=1);
+
 namespace VladimirYuldashev\Flysystem\Tests;
 
 use League\Flysystem\Config;
-use League\Flysystem\Util;
+use League\Flysystem\UnableToCopyFile;
+use League\Flysystem\UnableToRetrieveMetadata;
+use League\Flysystem\UnableToSetVisibility;
 
 class CurlFtpAdapterTest extends TestCase
 {
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $filename
      */
-    public function testWrite($filename): void
+    public function testWrite(string $filename): void
     {
         $contents = $this->faker()->text;
 
         $this->createResourceDirIfPathHasDir($filename);
 
-        $result = $this->adapter->write($filename, $contents, new Config);
+        $this->adapter->write($filename, $contents, self::publicConfig());
 
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $filename,
-            'contents' => $contents,
-            'mimetype' => Util::guessMimeType($this->getResourceAbsolutePath($filename), $contents),
-        ], $result);
+        $this->assertTrue($this->adapter->fileExists($filename));
 
         $this->assertEquals($contents, $this->getResourceContent($filename));
     }
 
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $filename
      */
-    public function testUpdate($filename): void
+    public function testUpdate(string $filename): void
     {
         $contents = $this->faker()->text;
 
         $this->createResourceDirIfPathHasDir($filename);
 
-        $this->adapter->write($filename, $contents, new Config);
+        $this->adapter->write($filename, $contents, self::publicConfig());
         $this->assertEquals($contents, $this->getResourceContent($filename));
 
         $newContents = $this->faker()->text;
-        $result = $this->adapter->update($filename, $newContents, new Config);
-
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $filename,
-            'contents' => $newContents,
-            'mimetype' => Util::guessMimeType($this->getResourceAbsolutePath($filename), $contents),
-        ], $result);
+        $this->adapter->write($filename, $newContents, self::publicConfig());
 
         $this->assertNotEquals($contents, $this->getResourceContent($filename));
         $this->assertEquals($newContents, $this->getResourceContent($filename));
@@ -60,10 +48,8 @@ class CurlFtpAdapterTest extends TestCase
 
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $filename
      */
-    public function testUpdateStream($filename): void
+    public function testUpdateStream(string $filename): void
     {
         $contents = $this->faker()->text;
 
@@ -73,7 +59,7 @@ class CurlFtpAdapterTest extends TestCase
 
         $this->createResourceDirIfPathHasDir($filename);
 
-        $this->adapter->writeStream($filename, $stream, new Config);
+        $this->adapter->writeStream($filename, $stream, self::publicConfig());
         $this->assertEquals($contents, $this->getResourceContent($filename));
 
         $newContents = $this->faker()->text;
@@ -82,7 +68,7 @@ class CurlFtpAdapterTest extends TestCase
         fwrite($stream, $newContents);
         rewind($stream);
 
-        $this->adapter->updateStream($filename, $stream, new Config);
+        $this->adapter->writeStream($filename, $stream, self::publicConfig());
 
         $this->assertNotEquals($contents, $this->getResourceContent($filename));
         $this->assertEquals($newContents, $this->getResourceContent($filename));
@@ -90,166 +76,157 @@ class CurlFtpAdapterTest extends TestCase
 
     /**
      * @dataProvider filesProvider
-     *
-     * @param $filename
      */
-    public function testRename($filename): void
+    public function testRename(string $filename): void
     {
         $this->adapter->write($filename, 'foo', new Config);
 
         $newFilename = $this->randomFileName();
 
-        $result = $this->adapter->rename($filename, $newFilename);
+        $this->adapter->move($filename, $newFilename, new Config);
 
-        $this->assertTrue($result);
-        $this->assertFalse($this->adapter->has($filename));
-        $this->assertNotFalse($this->adapter->has($newFilename));
+        $this->assertFalse($this->adapter->fileExists($filename));
+        $this->assertNotFalse($this->adapter->fileExists($newFilename));
     }
 
     /**
      * @dataProvider filesProvider
-     *
-     * @param $filename
      */
-    public function testCopy($filename): void
+    public function testCopy(string $filename): void
     {
         $this->adapter->write($filename, 'foo', new Config);
 
-        $result = $this->adapter->copy($filename, 'bar');
+        $this->adapter->copy($filename, 'bar', new Config);
 
-        $this->assertTrue($result);
-        $this->assertNotFalse($this->adapter->has($filename));
-        $this->assertNotFalse($this->adapter->has('bar'));
-        $this->assertEquals($this->adapter->read($filename)['contents'], $this->adapter->read('bar')['contents']);
+        $this->assertNotFalse($this->adapter->fileExists($filename));
+        $this->assertNotFalse($this->adapter->fileExists('bar'));
+        $this->assertEquals($this->adapter->read($filename), $this->adapter->read('bar'));
+    }
 
-        $this->assertFalse($this->adapter->copy('foo-bar', 'bar-foo'));
+    public function testCopyFails(): void
+    {
+        $this->expectException(UnableToCopyFile::class);
+        $this->adapter->copy('foo-bar', 'bar-foo', new Config);
     }
 
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $filename
      */
-    public function testDelete($filename): void
+    public function testDelete(string $filename): void
     {
         $this->createResourceDirIfPathHasDir($filename);
 
         $this->adapter->write($filename, 'foo', new Config);
 
-        $result = $this->adapter->delete($filename);
+        $this->adapter->delete($filename);
 
-        $this->assertTrue($result);
-        $this->assertFalse($this->adapter->has($filename));
+        $this->assertFalse($this->adapter->fileExists($filename));
     }
 
     public function testCreateAndDeleteDir(): void
     {
-        $result = $this->adapter->createDir('foo', new Config);
+        $this->assertFalse($this->adapter->directoryExists('foo'));
 
-        $this->assertSame(['type' => 'dir', 'path' => 'foo'], $result);
+        $this->adapter->createDirectory('foo', self::publicConfig());
 
-        $result = $this->adapter->deleteDir('foo');
+        $this->assertTrue($this->adapter->directoryExists('foo'));
 
-        $this->assertTrue($result);
+        $this->adapter->deleteDirectory('foo');
+
+        $this->assertFalse($this->adapter->directoryExists('foo'));
     }
 
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $filename
      */
-    public function testGetSetVisibility($filename): void
+    public function testGetSetVisibility(string $filename): void
     {
         $this->createResourceDirIfPathHasDir($filename);
 
         $this->adapter->write($filename, 'foo', new Config);
 
-        $result = $this->adapter->setVisibility($filename, 'public');
+        $this->adapter->setVisibility($filename, 'public');
 
-        $this->assertNotFalse($result);
-        $this->assertSame('public', $result['visibility']);
-        $this->assertSame('public', $this->adapter->getVisibility($filename)['visibility']);
+        $this->assertSame('public', $this->adapter->visibility($filename)->visibility());
 
-        $result = $this->adapter->setVisibility($filename, 'private');
+        $this->adapter->setVisibility($filename, 'private');
 
-        $this->assertNotFalse($result);
-        $this->assertSame('private', $result['visibility']);
-        $this->assertSame('private', $this->adapter->getVisibility($filename)['visibility']);
+        $this->assertSame('private', $this->adapter->visibility($filename)->visibility());
+    }
 
-        $this->assertFalse($this->adapter->setVisibility('bar', 'public'));
+    public function testGetSetVisibilityFails(): void
+    {
+        $this->expectException(UnableToSetVisibility::class);
+        $this->adapter->setVisibility('bar', 'public');
     }
 
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $name
      */
-    public function testRead($name): void
+    public function testRead(string $name): void
     {
         $contents = $this->faker()->text;
         $this->createResourceFile($name, $contents);
 
         $response = $this->adapter->read($name);
 
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $name,
-            'contents' => $contents,
-        ], $response);
-    }
-
-    public function testGetMetadata(): void
-    {
-        $this->assertSame(['type' => 'dir', 'path' => ''], $this->adapter->getMetadata(''));
+        $this->assertSame($contents, $response);
     }
 
     /**
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $name
      */
-    public function testHas($name): void
+    public function testFileExists(string $name): void
     {
         $contents = $this->faker()->text;
         $this->createResourceFile($name, $contents);
 
-        $this->assertTrue((bool) $this->adapter->has($name));
+        $this->assertTrue((bool) $this->adapter->fileExists($name));
     }
 
-    public function testGetMimeType(): void
+    public function testMimeType(): void
     {
         $this->adapter->write('foo.json', 'bar', new Config);
 
-        $this->assertSame('application/json', $this->adapter->getMimetype('foo.json')['mimetype']);
-        $this->assertFalse($this->adapter->getMimetype('bar.json'));
+        $this->assertSame('application/json', $this->adapter->mimeType('foo.json')->mimeType());
     }
 
-    public function testGetTimestamp(): void
+    public function testMimeTypeFaileNotExist(): void
     {
-        $this->assertFalse($this->adapter->getTimestamp('foo'));
+        $this->expectException(UnableToRetrieveMetadata::class);
+        $this->adapter->mimeType('bar.json');
+    }
+
+    public function testLastModifiedFailed(): void
+    {
+        $this->expectException(UnableToRetrieveMetadata::class);
+        $this->adapter->lastModified('foo');
+    }
+
+    public function testLastModified(): void
+    {
+        $this->adapter->write('foo.json.lastModified', 'bar', new Config);
+        $lastModified = $this->adapter->lastModified('foo.json.lastModified')->lastModified();
+        $this->assertGreaterThan(strtotime('yesterday 0:00'), $lastModified);
     }
 
     /**
      * @dataProvider withSubFolderProvider
-     *
-     * @param $path
      */
-    public function testListContents($path): void
+    public function testListContents(string $path): void
     {
         $contents = $this->faker()->text;
         $this->createResourceFile($path, $contents);
 
-        $this->assertCount(1, $this->adapter->listContents(dirname($path)));
+        $this->assertCount(1, iterator_to_array($this->adapter->listContents(dirname($path), false), false));
     }
 
     /**
      * @dataProvider withSubFolderProvider
-     *
-     * @param $path
      */
-    public function testListContentsEmptyPath($path): void
+    public function testListContentsEmptyPath(string $path): void
     {
-        $this->assertCount(0, $this->adapter->listContents(dirname($path)));
+        $this->assertCount(0, iterator_to_array($this->adapter->listContents(dirname($path), false), false));
     }
 
     /**
@@ -257,23 +234,17 @@ class CurlFtpAdapterTest extends TestCase
      * after reading a file, especially if this file is in a subfolder.
      *
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $path
      */
-    public function testReadAndHasInSequence($path): void
+    public function testReadAndHasInSequence(string $path): void
     {
         $contents = $this->faker()->text;
         $this->createResourceFile($path, $contents);
 
         $response = $this->adapter->read($path);
 
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $path,
-            'contents' => $contents,
-        ], $response);
+        $this->assertSame($contents, $response);
 
-        $this->assertTrue((bool) $this->adapter->has($path));
+        $this->assertTrue((bool) $this->adapter->fileExists($path));
     }
 
     /**
@@ -281,27 +252,18 @@ class CurlFtpAdapterTest extends TestCase
      * after writing a file, especially if this file is in a subfolder.
      *
      * @dataProvider filesAndSubfolderFilesProvider
-     *
-     * @param $path
      */
-    public function testWriteAndHasInSequence($path): void
+    public function testWriteAndHasInSequence(string $path): void
     {
         $contents = $this->faker()->text;
 
         $this->createResourceDirIfPathHasDir($path);
 
-        $result = $this->adapter->write($path, $contents, new Config);
-
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $path,
-            'contents' => $contents,
-            'mimetype' => Util::guessMimeType($this->getResourceAbsolutePath($path), $contents),
-        ], $result);
+        $this->adapter->write($path, $contents, self::publicConfig());
 
         $this->assertEquals($contents, $this->getResourceContent($path));
 
-        $this->assertTrue((bool) $this->adapter->has($path));
+        $this->assertTrue((bool) $this->adapter->fileExists($path));
     }
 
     /**
@@ -320,13 +282,9 @@ class CurlFtpAdapterTest extends TestCase
 
         $response = $this->adapter->read($read_path);
 
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $read_path,
-            'contents' => $read_path_contents,
-        ], $response);
+        $this->assertSame($read_path_contents, $response);
 
-        $this->assertTrue((bool) $this->adapter->has($has_path));
+        $this->assertTrue((bool) $this->adapter->fileExists($has_path));
     }
 
     /**
@@ -344,19 +302,12 @@ class CurlFtpAdapterTest extends TestCase
 
         $this->createResourceDirIfPathHasDir($write_path);
 
-        $response = $this->adapter->write($write_path, $write_path_contents, new Config);
+        $this->adapter->write($write_path, $write_path_contents, new Config);
 
-        $this->assertSame([
-            'type' => 'file',
-            'path' => $write_path,
-            'contents' => $write_path_contents,
-            'mimetype' => Util::guessMimeType($this->getResourceAbsolutePath($write_path), $write_path_contents),
-        ], $response);
-
-        $this->assertTrue((bool) $this->adapter->has($has_path));
+        $this->assertTrue((bool) $this->adapter->fileExists($has_path));
     }
 
-    public function filesAndSubfolderFilesProvider()
+    public static function filesAndSubfolderFilesProvider(): array
     {
         return [
             ['test.txt'],
@@ -364,25 +315,25 @@ class CurlFtpAdapterTest extends TestCase
             ['test 1.txt'],
             ['test  2.txt'],
             ['тест.txt'],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
             ['test/test.txt'],
             ['тёст/тёст.txt'],
             ['test 1/test.txt'],
             ['test/test 1.txt'],
             ['test  1/test  2.txt'],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
         ];
     }
 
-    public function filesProvider()
+    public static function filesProvider(): array
     {
         return [
             ['test.txt'],
@@ -390,15 +341,15 @@ class CurlFtpAdapterTest extends TestCase
             ['test 1.txt'],
             ['test  2.txt'],
             ['тест.txt'],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
-            [$this->randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
+            [self::randomFileName()],
         ];
     }
 
-    public function withSubFolderProvider()
+    public static function withSubFolderProvider(): array
     {
         return [
             ['test/test.txt'],
@@ -406,11 +357,11 @@ class CurlFtpAdapterTest extends TestCase
             ['test 1/test.txt'],
             ['test/test 1.txt'],
             ['test  1/test  2.txt'],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
-            [$this->faker()->word.'/'.$this->randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
+            [self::faker()->word.'/'.self::randomFileName()],
         ];
     }
 }
